test: Add test for when subcommand is not provided
